# fix #104: EntitySet performance slow for large number of entities