EntitySet uses new 'DictList' data structure for its internal _list
and _interestingEntities collections. Rough testing indicates Add,
Remove and Detach are now constant time (vs proportional to number
of entities in set, previously). See performance measurements below.

I couldn't get all tests to pass in master, I assume due to not
setting up test databases. There's one additional failing test in
this change, but I'm not sure it's my fault. For some reason,
WebAuthenticationServiceTest.SaveUserEmptyAsync is ignored in 
master, but runs & fails with these changes. I can't figure out
how it's getting ignored in master.

Note that I didn't use a simple HashSet, since for INotifyCollectionChanged,
EntitySet needs to support indexing of the collection. The order
of entities on DictList will be modified by Remove(), but I'm not
sure the order of entities in EntitySet is important. Since no
additional tests failed (apart from that unrelated (?) one), I
assume this is correct.


# EntitySet performance measurements (from temporary test in this branch)

Each line shows n: x where n is the number of entities, x is the time taken in ms.

master:						using DictList:
	add						add
	400: 	610				400: 187
	800: 	340				800: 143
	1600: 	644				1600: 259
	3200: 	1261			3200: 765
	6400: 	3594			6400: 1222
	12800: 	12119			12800: 2275
	attach					attach
	400: 	77				400: 80
	800: 	150				800: 137
	1600: 	308				1600: 265
	3200: 	592				3200: 560
	6400: 	1284			6400: 1157
	12800: 	2417			12800: 2239
	modify 	all				modify all
	800: 	46				800: 47
	1600: 	106				1600: 87
	3200: 	204				3200: 218
	6400: 	436				6400: 437
	12800: 	812				12800: 913
	25600: 	1837			25600: 1528
	remove 	all				remove all
	800: 	13				800: 7
	1600: 	45				1600: 6
	3200: 	113				3200: 11
	6400: 	602				6400: 32
	12800: 	2703			12800: 60
	25600: 	21500			25600: 143
	detach 	all				detach all
	800: 	16				800: 26
	1600: 	32				1600: 58
	3200: 	75				3200: 63
	6400: 	283				6400: 127
	12800: 	835				12800: 297
	25600: 	2967			25600: 583