@Daniel-Svensson 2018-01-21
Good job, that looks promising!
I am currently away for a few weeks but once home I will look into this, I need to get in contact with outercurve to see how we should handle cla signing as well as loom through the code.

As for the test not all of them pass (around 26 fails when running on vsts and more on appveyor). You must setup the database (setup-database.ps1) and make sure the webserver is running the "fulltrust project"and even then a few are flaky and requires a rerun.