@Daniel-Svensson 2018-01-31
I've created a simple benchmark to help you at https://gist.github.com/Daniel-Svensson/c0dabcb6c4dd8ccab06dedb9923c81a6 

Just create a new project and add a nuget reference to BenchmarkDotNet and project references to OpenRiaServices.DomainServices.Client and OpenRiaServices.DomainServices.Client.Test projects.

### Some output:

The code is not against master but with some other minor perf improvements included (https://github.com/Daniel-Svensson/OpenRiaServices-1/tree/feature/metamember_performance) so your measurements will be slightly different. 
The key takeaway is most operations apart from Attach (which has regressed slightly) has pretty solid performance improvements.

#### Without changes:

Add & Attach have had the "None" measurements subtracted
Benchmarks including Add have had the Add measurements subtracted.

|Method|NumEntities|Mean (us)|Gen 0|Allocated (KB)|
|--|--|--|--|--|
|Add|10|303,3|33,6914|103,96|
|Attach|10|336|40,0391|123,48|
|AddAndModify|10|440,2|50,293|155,75|
|AddAndRemove|10|312,7|34,1797|106,39|
|AddAndDetach|10|337,1|41,0156|127,29|
|Add|500|19746,4|843,75|5151,76|
|Attach|500|19456|1000|6137,98|
|AddAndModify|500|26264,9|1250|7739,8|
|AddAndRemove|500|21164,1|843,75|5269,15|
|AddAndDetach|500|21108,8|1031,25|6314,02|
|Add|5000|438250,7|9875|51629,97|
|Attach|5000|218924,7|11250|61459,1|
|AddAndModify|5000|529264,3|14437,5|77517,25|
|AddAndRemove|5000|538455,5|9687,5|52786,91|
|AddAndDetach|5000|478795,7|11812,5|63256,06|

#### With changes

All benchmarks have had the "None" measurements removed
Benchmarks including Add have not had the add measurements removed.

|Method|NumEntities|Mean us|Gen 0|Allocated (KB)|
|--|--|--|--|--|
|Add|10|297,8|34,1797|105,13|
|Attach|10|328,2|40,0391|124,23|
|AddAndModify|10|415,8|50,7813|156,92|
|AddAndRemove|10|308|34,668|107,56|
|AddAndDetach|10|329,4|41,5039|128,46|
|Add|500|17242,2|843,75|5210,84|
|Attach|500|19866,6|1000|6171,56|
|AddAndModify|500|24106,1|1281,25|7798,93|
|AddAndRemove|500|17674|875|5328,36|
|AddAndDetach|500|18516,4|1031,25|6373,11|
|Add|5000|198007,9|9875|52182,51|
|Attach|5000|221340|10937,5|61762,69|
|AddAndModify|5000|282519|14500|78069,61|
|AddAndRemove|5000|201566,3|10062,5|53354,29|
|AddAndDetach|5000|224142,5|11812,5|63809,11|