@Daniel-Svensson 2018-02-04
Take your time, I have tried to contact outercurve as well as the dotnet foundation since outercurve does not seems to exist anymore and they seems to have dropped their DNS domain.

My guess is that it can take weeks to have a CLA process in place so that I feel confident to accept larger contributions. In the mean time I am considering to release a new minor release (at least as beta) with your other performance improvements.