@uozuAho 2018-02-10
Hi Daniel. I've addressed all the changes above, however performance for Remove and Detach is now worse than in master. I may have done something wrong there, so I'll look into it a bit more. It may not be possible to do better using the current IndexedList data structure, since it needs to traverse the list and its index for every Remove operation.