@uozuAho 2018-03-24
Sorry for the lack of recent activity on this, I had a look into #113 since it blocks this PR. Running tests and benchmarks is pretty time consuming, so this has been on the backburner :(