@Daniel-Svensson 2018-03-27
Take your time, there is no hurry. 
I've tried to Contact the dotnetfoundation again, this time by email last weekend in order to see if they can give some advice on licensing / CLA so I can feel Confident merging larger contributions.

I've updated #113 with some notes that might help.
If you get stuck and could use some assistance please let me know.

Keep up the god job :+1: your efforts are appreciated.