@uozuAho 2018-05-16
Hi Daniel, I'm thinking of abandoning this attempt. #126 achieves some good gains with fewer code changes. I don't think the data structure I'm using for this PR is appropriate - benchmarks are showing that Remove is 2-3x slower for 5000 entities. Some googling tells me that it may be possible to achieve log(n) performance for add/contains/remove/index with a tree structure, but the code would be complicated, and even then maybe the performance hit to add & index may not be worth it.

Thanks for taking the time to review this code (multiple times!).